# Switch HLC timestamps to be stored as an array of two int64s