Benchmarks:

```
benchmark                                                                                     old ns/op     new ns/op     delta
BenchmarkHLCParsing/1-10                                                                      122           44.2          -63.64%
BenchmarkHLCParsing/2-10                                                                      121           44.6          -63.17%
BenchmarkHLCParsing/42-10                                                                     128           48.5          -62.24%
BenchmarkHLCParsing/1257894000000000000-10                                                    248           71.7          -71.10%
BenchmarkHLCParsing/-1-10                                                                     122           46.6          -61.87%
BenchmarkHLCParsing/9223372036854775807.1000000025-10                                         376           103           -72.71%
BenchmarkHLCParsing/1703283409994227985.0000000004-10                                         376           102           -72.81%
BenchmarkHLCLessThan/1-2-10                                                                   0.00          0.00          -100.00%
BenchmarkHLCLessThan/2-1-10                                                                   0.00          0.00          +0.00%
BenchmarkHLCLessThan/2-2-10                                                                   0.00          0.00          +0.00%
BenchmarkHLCLessThan/1703283409994227985.0000000004-1703283409994227985.0000000005-10         0.00          0.00          -80.00%
BenchmarkHLCLessThan/1703283409994227985.0000000005-1703283409994227985.0000000004-10         0.00          0.00          -100.00%
BenchmarkHLCLessThanFunc/1-2-10                                                               242           0.33          -99.86%
BenchmarkHLCLessThanFunc/2-1-10                                                               240           0.33          -99.86%
BenchmarkHLCLessThanFunc/2-2-10                                                               243           0.32          -99.87%
BenchmarkHLCLessThanFunc/1703283409994227985.0000000001-1703283409994227985.0000000001-10     750           0.32          -99.96%
BenchmarkHLCLessThanFunc/1703283409994227985.0000000004-1703283409994227985.0000000005-10     757           0.33          -99.96%
BenchmarkHLCLessThanFunc/1703283409994227985.0000000005-1703283409994227985.0000000004-10     751           0.33          -99.96%
```